pyproject.toml
[tool.poetry]
name = "flox-gitlab"
version = "0.1.0"
description = ""
authors = ["Daniel Ancuta <whisller@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
flox-core = {path = "../flox-core"}
python-gitlab = "^2.1.2"

[tool.poetry.dev-dependencies]

[tool.poetry.plugins."flox.plugin"]
gitlab = "flox_gitlab:plugin"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
